+++ C++/test/algorithm/sorting/sorting.cpp
@@ -7,6 +7,7 @@
 #include "algorithm/sorting/insertion_sort.hpp"
 #include "algorithm/sorting/merge_sort.hpp"
 #include "algorithm/sorting/quick_sort.hpp"
+#include "algorithm/sorting/selection_sort.hpp"
 
 // Prototypes
 int generate_random_int(int, int);
@@ -26,7 +27,8 @@ TEST_CASE("Sort in ascending order", "[sorting]") {
 			heap_sort,
 			insertion_sort,
 			merge_sort,
-			quick_sort
+			quick_sort,
+			selection_sort
 	};
 
 	vector<int> original, algo_sorted, std_sorted;
@@ -53,7 +55,8 @@ TEST_CASE("Sort in descending order", "[sorting]") {
 			heap_sort,
 			insertion_sort,
 			merge_sort,
-			quick_sort
+			quick_sort,
+			selection_sort
 	};
 
 	vector<int> original, algo_sorted, std_sorted;
